faiss-service/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.5.1+cpu
faiss-cpu
fastapi
uvicorn
numpy
sentence-transformers
